refactoring tests structure
