+++ microservices/recommendation-service/src/test/java/com/yavlash/microservices/core/recommendation/repository/MongoDbTestBase.java
@@ -1,4 +1,4 @@
-package com.yavlash.microservices.core.recommendation;
+package com.yavlash.microservices.core.recommendation.repository;
 
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
